requirements.txt
playwright>=1.40.0
pyyaml>=6.0.1
requests>=2.31.0
openai>=1.3.0
apscheduler>=3.10.4
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
cssutils>=2.9.0
lxml>=4.9.0
watchdog>=3.0.0
Pygments>=2.15.0
